# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "This is a Audit log swagger server.",
        "title": "Audit log Swagger API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/logs": {
            "get": {
                "description": "Get a list of audit logs with filtering options",
                "tags": [
                    "audit_logs"
                ],
                "summary": "List audit logs",
                "parameters": [
                    {
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Filter by user ID",
                        "name": "user_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by action",
                        "name": "action",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by resource type",
                        "name": "resource_type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by severity",
                        "name": "severity",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by start time (RFC3339)",
                        "name": "start_time",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by end time (RFC3339)",
                        "name": "end_time",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/domain.AuditLog"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new audit log entry",
                "tags": [
                    "audit_logs"
                ],
                "summary": "Create audit log",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/domain.AuditLog"
                            }
                        }
                    },
                    "description": "Audit log object",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/domain.AuditLog"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logs/bulk": {
            "post": {
                "description": "Create multiple audit log entries in a single request",
                "tags": [
                    "audit_logs"
                ],
                "summary": "Bulk create audit logs",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/domain.AuditLog"
                                }
                            }
                        }
                    },
                    "description": "Array of audit log objects",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/domain.AuditLog"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logs/cleanup": {
            "delete": {
                "description": "Delete audit logs older than specified days",
                "tags": [
                    "audit_logs"
                ],
                "summary": "Delete old logs",
                "parameters": [
                    {
                        "description": "Number of days to retain",
                        "name": "days",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logs/export": {
            "get": {
                "description": "Export audit logs with filtering options in JSON or CSV format",
                "tags": [
                    "audit_logs"
                ],
                "summary": "Export audit logs",
                "parameters": [
                    {
                        "description": "Export format (json or csv)",
                        "name": "format",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "json"
                        }
                    },
                    {
                        "description": "Filter by user ID",
                        "name": "user_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by action",
                        "name": "action",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by resource type",
                        "name": "resource_type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by severity",
                        "name": "severity",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by start time (RFC3339)",
                        "name": "start_time",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by end time (RFC3339)",
                        "name": "end_time",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            },
                            "text/csv": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            },
                            "text/csv": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            },
                            "text/csv": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            },
                            "text/csv": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logs/stats": {
            "get": {
                "description": "Get statistics about audit logs including counts by action, severity, and resource",
                "tags": [
                    "audit_logs"
                ],
                "summary": "Get log statistics",
                "parameters": [
                    {
                        "description": "Filter by start time (RFC3339)",
                        "name": "start_time",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by end time (RFC3339)",
                        "name": "end_time",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/domain.AuditLogStats"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logs/{id}": {
            "get": {
                "description": "Get an audit log entry by its ID",
                "tags": [
                    "audit_logs"
                ],
                "summary": "Get audit log",
                "parameters": [
                    {
                        "description": "Log ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/domain.AuditLog"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "servers": [
        {
            "url": "//localhost:10000/api/v1"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
            "domain.AuditLog": {
                "type": "object"
            },
            "domain.AuditLogStats": {
                "type": "object",
                "properties": {
                    "action_counts": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "integer"
                        }
                    },
                    "resource_counts": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "integer"
                        }
                    },
                    "severity_counts": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "integer"
                        }
                    },
                    "total_logs": {
                        "type": "integer"
                    }
                }
            },
...
            "dto.Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "error message"
                    }
                }
            }
        }
    }
}